+++ DX11/2DFrameWork/ObTileMap.h
@@ -52,7 +52,7 @@ public:
     void        Render()override;
     void        SetTile(Int2 TileIdx, Int2 FrameIdx, int ImgIdx = 0 ,
         byte TileState = TILE_NONE , Color color = Color(0.5f, 0.5f, 0.5f, 0.5f));
-    void        SetLight(Int2 TileIdx, byte light);
+    void        SetLight(Int2 TileIdx, int light);
     void        UpdateSub();
     int         GetTileState(Int2 TileIdx);
     void        SetTileState(Int2 TileIdx, int TileState);
