+++ DX11/Game1/Tilebuild.h
@@ -3,9 +3,9 @@
 class Tilebuild
 {
 public:
-	void DirtToGrass(ObTileMap& block, Int2 pos, byte** mapLight);
-	void TileRemove(ObTileMap& block, Int2 tilePos, Map& map, byte** mapLight, bool** wall);
-	void TileAdd(ObTileMap& block, Int2 tilePos, Map& map, byte type, byte** mapLight);
-	void TileArrangement(ObTileMap& block, Int2 pos, byte type, Map& map, byte** mapLight);
+	void DirtToGrass(ObTileMap& block, Int2 pos, int** mapLight);
+	void TileRemove(ObTileMap& block, Int2 tilePos, Map& map, int** mapLight, bool** wall);
+	void TileAdd(ObTileMap& block, Int2 tilePos, Map& map, byte type, int** mapLight);
+	void TileArrangement(ObTileMap& block, Int2 pos, byte type, Map& map, int** mapLight);
 };
